src/cv/data/autodistill-video-project/pyproject.toml
[project]
name = "autodistill-video-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autodistill>=0.1.29",
    "autodistill-grounded-sam>=0.1.2",
    "autodistill-grounding-dino>=0.1.4",
    "numpy>=2.4.2",
    "opencv-python-headless>=4.13.0.92",
    "roboflow>=1.1.4",
    "scikit-learn>=1.8.0",
    "setuptools>=80.10.2",
    "supervision>=0.27.0.post1",
    "tqdm>=4.67.3",
    "transformers<5.0.0",
    "ultralytics>=8.3.0",
    "pyyaml>=6.0.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
]
